+++ src/main/java/org/folio/search/cql/CqlSearchQueryConverter.java
@@ -3,6 +3,8 @@ package org.folio.search.cql;
 import static org.folio.search.utils.SearchQueryUtils.isBoolQuery;
 import static org.folio.search.utils.SearchQueryUtils.isDisjunctionFilterQuery;
 import static org.folio.search.utils.SearchQueryUtils.isFilterQuery;
+import static org.folio.search.utils.SearchUtils.SHARED_FIELD_NAME;
+import static org.folio.search.utils.SearchUtils.TENANT_ID_FIELD_NAME;
 import static org.opensearch.index.query.QueryBuilders.boolQuery;
 
 import java.util.ArrayList;
@@ -61,6 +63,7 @@ public class CqlSearchQueryConverter {
 
   //todo(MSEARCH-576): may be reworked after implemented for browse/streamIds.
   // Implemented separately because it crashes 'browse/streamIds' functionality.
+
   /**
    * Converts given CQL search query value to the elasticsearch {@link SearchSourceBuilder} object.
    * Wraps base 'convert' and adds tenantId+shared filter in case of consortia mode
@@ -146,6 +149,61 @@ public class CqlSearchQueryConverter {
     conditions.add(leftOperandQuery);
     conditions.add(rightOperandQuery);
     return boolQuery;
+  }
+
+  private QueryBuilder filterForActiveAffiliation(QueryBuilder query) {
+    var contextTenantId = folioExecutionContext.getTenantId();
+    var centralTenantId = consortiumTenantService.getCentralTenant(contextTenantId);
+    if (centralTenantId.isEmpty()) {
+      return query;
+    }
+
+    var boolQuery = prepareBoolQueryForActiveAffiliation(query);
+    addActiveAffiliationClauses(boolQuery, contextTenantId, centralTenantId.get());
+
+    return boolQuery;
+  }
+
+  private BoolQueryBuilder prepareBoolQueryForActiveAffiliation(QueryBuilder query) {
+    BoolQueryBuilder boolQuery;
+    if (query instanceof MatchAllQueryBuilder) {
+      boolQuery = boolQuery();
+    } else if (query instanceof BoolQueryBuilder bq) {
+      boolQuery = bq;
+    } else {
+      boolQuery = boolQuery().must(query);
+    }
+    boolQuery.minimumShouldMatch(1);
+    return boolQuery;
+  }
+
+  private void addActiveAffiliationClauses(BoolQueryBuilder boolQuery, String contextTenantId, String centralTenantId) {
+    var affiliationShouldClauses = getAffiliationShouldClauses(contextTenantId, centralTenantId);
+    if (boolQuery.should().isEmpty()) {
+      affiliationShouldClauses.forEach(boolQuery::should);
+    } else {
+      var innerBoolQuery = boolQuery();
+      affiliationShouldClauses.forEach(innerBoolQuery::should);
+      boolQuery.must(innerBoolQuery);
+    }
+  }
+
+  private LinkedList<QueryBuilder> getAffiliationShouldClauses(String contextTenantId, String centralTenantId) {
+    var affiliationShouldClauses = new LinkedList<QueryBuilder>();
+    addTenantIdAffiliationShouldClause(contextTenantId, centralTenantId, affiliationShouldClauses);
+    addSharedAffiliationShouldClause(affiliationShouldClauses);
+    return affiliationShouldClauses;
+  }
+
+  private void addTenantIdAffiliationShouldClause(String contextTenantId, String centralTenantId,
+                                                  LinkedList<QueryBuilder> affiliationShouldClauses) {
+    if (!contextTenantId.equals(centralTenantId)) {
+      affiliationShouldClauses.add(termQuery(TENANT_ID_FIELD_NAME, contextTenantId));
+    }
+  }
+
+  private void addSharedAffiliationShouldClause(LinkedList<QueryBuilder> affiliationShouldClauses) {
+    affiliationShouldClauses.add(termQuery(SHARED_FIELD_NAME, true));
   }
 
   private QueryBuilder enhanceQuery(QueryBuilder query, String resource) {